refactor(contracts): move contracts package to interfaces layer
Move bioetl.contracts to bioetl.interfaces.contracts per user request.
The interfaces layer is the appropriate location for data contracts
as they define the external API for data consumers (analysts).

New canonical import path:
    from bioetl.interfaces.contracts import ChEMBLActivityGoldSchema

Backward compatibility maintained via re-exports in:
- bioetl.infrastructure.schemas.gold (deprecated)
