+++ src/bioetl/infrastructure/schemas/gold.py
@@ -1,11 +1,11 @@
 """Gold layer Pandera schemas for data validation.
 
-DEPRECATED: This module re-exports schemas from bioetl.contracts for
-backward compatibility. New code should import from bioetl.contracts directly.
+DEPRECATED: This module re-exports schemas from bioetl.interfaces.contracts for
+backward compatibility. New code should import from bioetl.interfaces.contracts.
 
 Example:
     # Preferred (new code):
-    from bioetl.contracts import ChEMBLActivityGoldSchema
+    from bioetl.interfaces.contracts import ChEMBLActivityGoldSchema
 
     # Deprecated (legacy code, still works):
     from bioetl.infrastructure.schemas.gold import ChEMBLActivityGoldSchema
@@ -14,7 +14,7 @@ Example:
 from __future__ import annotations
 
 # Re-export all schemas from contracts for backward compatibility
-from bioetl.contracts.gold import (
+from bioetl.interfaces.contracts.gold import (
     ChEMBLActivityGoldSchema,
     ChEMBLAssayGoldSchema,
     ChEMBLAssayParametersGoldSchema,
@@ -35,7 +35,7 @@ from bioetl.contracts.gold import (
     UniProtIDMappingGoldSchema,
     UniProtProteinGoldSchema,
 )
-from bioetl.contracts.gold._base import DATE_REGEX
+from bioetl.interfaces.contracts.gold._base import DATE_REGEX
 
 __all__ = [
     "DATE_REGEX",
