+++ src/bioetl/interfaces/contracts/gold/__init__.py
@@ -1,38 +1,27 @@
-"""Data contracts for BioETL Gold layer.
+"""Gold layer data contracts organized by provider.
 
-This package provides Pandera DataFrameModel schemas for Gold layer validation.
-Schemas are independent of pipeline implementations and can be imported by
-data consumers (analysts, downstream applications) for validation and documentation.
+This package contains Pandera DataFrameModel schemas for Gold layer validation,
+organized by data provider for easy navigation.
 
-Usage for analysts:
-    >>> from bioetl.contracts import ChEMBLActivityGoldSchema
-    >>> import pandas as pd
+Submodules:
+    chembl: ChEMBL bioactivity database schemas
+    pubchem: PubChem compound schemas
+    uniprot: UniProt protein database schemas
+    publications: Cross-provider publication schemas (PubMed, CrossRef, OpenAlex, SemanticScholar)
+
+Example usage:
+    >>> from bioetl.interfaces.contracts.gold import chembl
     >>> df = pd.read_parquet("data/gold/chembl_activity/")
-    >>> ChEMBLActivityGoldSchema.validate(df)
-
-    # Or import by provider:
-    >>> from bioetl.contracts.gold import chembl
     >>> chembl.ChEMBLActivityGoldSchema.validate(df)
 
-Available schemas by provider:
-    - ChEMBL: Activity, Assay, Target, Molecule, etc.
-    - PubChem: Compound
-    - UniProt: Protein, IDMapping
-    - PubMed: Publication
-    - CrossRef: Publication
-    - OpenAlex: Publication
-    - SemanticScholar: Publication
-
-See also:
-    - docs/03-data-contracts/ for detailed schema documentation
-    - ADR-018 for Gold strict validation rationale
+    >>> from bioetl.interfaces.contracts.gold.publications import PubMedPublicationGoldSchema
+    >>> PubMedPublicationGoldSchema.validate(pubmed_df)
 """
 
 from __future__ import annotations
 
-# Re-export all Gold schemas for convenient access
-from bioetl.contracts.gold import (
-    # ChEMBL schemas
+# Import all schemas for flat namespace access
+from bioetl.interfaces.contracts.gold.chembl import (
     ChEMBLActivityGoldSchema,
     ChEMBLAssayGoldSchema,
     ChEMBLAssayParametersGoldSchema,
@@ -45,28 +34,20 @@ from bioetl.contracts.gold import (
     ChEMBLProteinClassGoldSchema,
     ChEMBLTargetComponentGoldSchema,
     ChEMBLTargetGoldSchema,
-    # CrossRef schemas
+)
+from bioetl.interfaces.contracts.gold.pubchem import PubChemCompoundGoldSchema
+from bioetl.interfaces.contracts.gold.publications import (
     CrossRefPublicationGoldSchema,
-    # OpenAlex schemas
     OpenAlexPublicationGoldSchema,
-    # PubChem schemas
-    PubChemCompoundGoldSchema,
-    # PubMed schemas
     PubMedPublicationGoldSchema,
-    # SemanticScholar schemas
     SemanticScholarPublicationGoldSchema,
-    # UniProt schemas
+)
+from bioetl.interfaces.contracts.gold.uniprot import (
     UniProtIDMappingGoldSchema,
     UniProtProteinGoldSchema,
 )
 
-# Utilities
-from bioetl.contracts.gold._base import DATE_REGEX
-
 __all__ = [
-    # Utilities
-    "DATE_REGEX",
-    # ChEMBL
     "ChEMBLActivityGoldSchema",
     "ChEMBLAssayGoldSchema",
     "ChEMBLAssayParametersGoldSchema",
@@ -79,17 +60,11 @@ __all__ = [
     "ChEMBLProteinClassGoldSchema",
     "ChEMBLTargetComponentGoldSchema",
     "ChEMBLTargetGoldSchema",
-    # CrossRef
     "CrossRefPublicationGoldSchema",
-    # OpenAlex
     "OpenAlexPublicationGoldSchema",
-    # PubChem
     "PubChemCompoundGoldSchema",
-    # PubMed
     "PubMedPublicationGoldSchema",
-    # SemanticScholar
     "SemanticScholarPublicationGoldSchema",
-    # UniProt
     "UniProtIDMappingGoldSchema",
     "UniProtProteinGoldSchema",
 ]
