+++ src/bioetl/interfaces/contracts/gold/_base.py
