+++ src/bioetl/interfaces/contracts/gold/chembl.py
@@ -25,7 +25,7 @@ from __future__ import annotations
 import pandera.pandas as pa
 from pandera.typing import Series
 
-from bioetl.contracts.gold._base import DATE_REGEX
+from bioetl.interfaces.contracts.gold._base import DATE_REGEX
 
 
 class ChEMBLActivityGoldSchema(pa.DataFrameModel):
