+++ src/bioetl/interfaces/contracts/gold/pubchem.py
