+++ src/bioetl/interfaces/contracts/gold/uniprot.py
