+++ src/tools/scripts/generate_contracts.py
@@ -13,7 +13,7 @@ project_root = Path(__file__).resolve().parent.parent
 sys.path.append(str(project_root / "src"))
 
 try:
-    from bioetl.contracts import (
+    from bioetl.interfaces.contracts import (
         ChEMBLActivityGoldSchema,
         PubChemCompoundGoldSchema,
         PubMedPublicationGoldSchema,
