+++ src/tools/verify_schema_parity.py
@@ -4,12 +4,6 @@ import dataclasses
 import logging
 import sys
 
-from bioetl.contracts import (
-    ChEMBLActivityGoldSchema,
-    ChEMBLAssayGoldSchema,
-    ChEMBLMoleculeGoldSchema,
-    ChEMBLTargetGoldSchema,
-)
 from bioetl.domain.entities.bioactivity import Bioactivity
 from bioetl.domain.entities.chembl_activity import Assay
 from bioetl.domain.entities.chembl_structures import Molecule, Target
@@ -18,6 +12,12 @@ from bioetl.infrastructure.schemas.silver import (
     CHEMBL_ASSAY_SCHEMA,
     CHEMBL_MOLECULE_SCHEMA,
     CHEMBL_TARGET_SCHEMA,
+)
+from bioetl.interfaces.contracts import (
+    ChEMBLActivityGoldSchema,
+    ChEMBLAssayGoldSchema,
+    ChEMBLMoleculeGoldSchema,
+    ChEMBLTargetGoldSchema,
 )
 
 # Configure logging for CLI output
